Update service_interface.go
client_interface and service_interface together display the basic functioning of the http service provided by Go. The client_interface runs on the user machine and the service_interface runs on the server. For now, no DB has been added to this code. Hence whenever a new auction request is made, it is done by submitting the http request in the browser directly. Once the service_interface is running, the item is registered on the server by submitting a request in the browser, the client_interface is run on the machine.
Description: A simple auction system, where an item added to the auction list via server request and the user can bid on the items or check the items present in the auction list.