Add files via upload
Account_pay.go>> This shows the use of locks and mutex in Golang. A simple pay service system is created to support client payments and account balance checking, users are already registered using Structs and maps are used to key a person. A person can check the balance, pay to another person, update the balance records.